requirements.txt
aspose-3d==24.9.0
attrs==23.2.0
backcall==0.2.0
Bottleneck
certifi 
cffi==1.15.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
cycler==0.11.0
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
fastjsonschema==2.19.1
Flask==2.2.5
fonttools==4.38.0
fvcore==0.1.5.post20210915
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.9
ipython==7.34.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.17.3
jupyter_core==4.12.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.4
lpips==0.1.4
MarkupSafe==2.1.5
mathutils==2.81.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1 
mkl-service==2.4.0
nbformat==5.7.0
nest-asyncio==1.6.0
networkx==2.6.3
numexpr
open3d==0.16.0
opencv-contrib-python==4.6.0.66
packaging==23.1
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
plotly==5.18.0
plyfile==0.9
portalocker==1.4.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.17.2
pyheif==0.7.1
pyparsing==3.0.9
pypng==0.20220715.0
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil
pytorch-msssim==1.0.0
pytz 
PyWavelets==1.3.0
PyYAML==6.0.1
retrying==1.3.4
Rtree==1.0.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tqdm==4.66.2
traitlets==5.9.0
trimesh==3.21.5
torchmetrics==1.5.2
wcwidth==0.2.13
Werkzeug==2.2.3
widgetsnbextension==4.0.10
yacs==0.1.8
zipp==3.15.0
tinycudann @ git+https://github.com/NVlabs/tiny-cuda-nn/@2ec562e853e6f482b5d09168705205f46358fb39#subdirectory=bindings/torch
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@47d5dc88247035b35ca3cfce159565f92d8fbb75
